requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
PyPDF2==3.0.1
pdf2image==1.16.3
plotly==5.18.0
matplotlib==3.8.2
wordcloud==1.9.3
Pillow==10.2.0
python-dotenv==1.0.0
numpy==1.26.3
pandas==2.1.4
